+++ changelog.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - MapSeries Print: Works with WebMercator (and different display spatial reference)
   [discusson #401]()https://github.com/e-netze/webgis-community/discussions/401
 
+- Undo Button: not (always) works in Mobile-Editing mode
+  [discusson #403]()https://github.com/e-netze/webgis-community/discussions/403
+
 ## 8.26.802
 
 ## Added